requirements.txt
Django==1.10.5
django-crispy-forms==1.6.1
django-formtools==2.0
django-jinja==2.3.1
django-reversion==2.0.9
djangorestframework==3.6.3
future==0.16.0
httplib2==0.10.3
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
pytz==2017.2
six==1.10.0
